Updated solution from array to stack
In order to be able to work with higher number of workshops I changed the data structure of the solution vector to a stack. This way it only ever has to have n+1 entries instead of 2^n.

The errors have gone away, but now I'm getting timed out on hackerrank.