Changed approach to sorting by end times
From discussion page I got a hint to think about the problem focusing on ending times. From there I realized that if there is an interval ending by 1 then you always take it, if not you look for an interval ending in 2. You repeat this, always taking the lowest interval available in this way because if you pick an interval ending at 2, and say you then find an interval from 1 to 3 which overlaps with that 2 interval it doesnt matter. Either there is an interval from 2 to 3 that you pick, or the choice between the 0 to 2 and 1 to 3 interval are arbitrary.

Its a neat solution, but it still timed out on 4/15 test cases.... have to see if I can optimize it further.